MB-7369 MB-7370 Add "instance" vclock to important per-node keys.
The motivation is basically the same as in
http://review.couchbase.org/21999. But there's one important detail
that was missed before. During conflict resolution two conflicting
vclocks are merged to the vclock that descends both conflicting
vclocks. It means that if the node with faked vclock is removed from
the cluster and then added back, our trick will not work this
time. Because merged vlcock will descend our faked vclock. And the new
value will get overwritten. So the solution is to attach "instance"
vclock to such vulnerable keys during config initialization. So that
it appears that those keys have been modified by some random
node. This will cause conflict during config merge and the newer value
will be preferred.

Change-Id: I7977ada2f8aed2e13374aa2ea72deb5796fc3eb2
Reviewed-on: http://review.couchbase.org/23376
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>