MB-7511 Handle pre-2.0.1 nodes when (un)inhibiting compaction.
We should not assume that there's previous compaction inhibition when
asked to compact a node. This is because this node may not support
compaction inhibition. And thus when we asked it to inhibit compaction
it replied with something that we treated as a 'nack'. And therefore,
compaction_inhibitions table doesn't have corresponding record.

Change-Id: I13505ab22717671ab631f8cc1e9d82820651c787
Reviewed-on: http://review.couchbase.org/23965
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>