MB-7574 Unsupport /polls/default/stats REST call.
This REST call was meant to aggregate stats for several buckets. And
it used to do so long time ago (Northscale Server 1.0.3). But after
'membase' bucket type was introduced, it worked only for a single
bucket and failed with badmatch error otherwise. The proper way to
grab bucket stats now is /pools/default/buckets/<bucket name>/stats
REST call.

Change-Id: Ia6f0360f5e5872feacdb2873e040e3334b6bc42c
Reviewed-on: http://review.couchbase.org/24115
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>