MB-7568 cas_remote_clusters handles lack of key in dynamic conf.
The culprit is that ns_config:update_key does not try to update keys
that are not present in dynamic config even if they are present in
default config. When 2.0 node joins 1.8.* cluster, its dynamic config
gets purged. Thus remote_clusters key is present only in default
config and cas_remote_clusters will just fail. To handle this we
introduce a version of ns_config:update_key that assigns default value
to a key when it's not found. This version is used only by
cas_remote_clusters, so otherwise behavior is unchanged.

A workaround for 2.0 is to restart the nodes. After this, everything
from the default config will be propagated to a dynamic config and
cas_remote_clusters will be able to succeed.

Change-Id: I5dd306c8753c55c19f974f3c572058cd182e07dc
Reviewed-on: http://review.couchbase.org/24143
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>