MB-7731 Count moves from 'undefined' nodes in total_in_flight.
If moves schedule contains only moves from 'undefined' nodes, we'll
immediately decide that rebalance is done. This is because 'undefined'
moves are not counted in total_in_flight. We don't expect that in
ns_vbucket_mover:on_backfill_done which causes badmatch error. Which
in turn causes rebalance to fail.

Change-Id: I5b1a31b5f51b99bf5ea2e0bf4a2a93ca0a421bc2
Reviewed-on: http://review.couchbase.org/24569
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Jin Lim <jin@couchbase.com>