MB-7432: add XDCR percentage of completeness stat
Yet another new stat required by customer and supporting engineers,
probably the last one under this bug. This stat tracks the progress
of XDCR by computing the percentage of work done out of all. Note
it is computed as the number of items checked, instead of items that
have been replicated, divided by all items (items checked plus to-
be-replicated). That means, if source and destination clusters
are already synced, nothing will be replicated and the completeness
will be 100%.

Change-Id: I2d284bf8e1da8e0001f6c6778abc098e59e1be05
Reviewed-on: http://review.couchbase.org/24583
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>