MB-7824 Convert messages to binary before sending to sinks.
Disk sinks do this anyway. So it will save us some work there and
unneeded copies between processes. Besides this, we now that logging
long strings takes a lot of space and erlang is quite reluctant about
garbage collecting it. So it's better to get rid of strings before
message reaches long-living sink processes. The caller is more likely
to be short-living process. So it's likely that memory will be
released sooner.

Change-Id: Ib2fb805194fb46ed3a09284f2cfe289266a44531
Reviewed-on: http://review.couchbase.org/24909
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>