MB-7824 Hibernate ns_log and ns_log_sink.
ns_log expects messages to be strings. And it's not easy to change it
to work with binaries because of backward compatibility issues. So
instead we just hibernate ns_log and ns_log_sink. Since we don't frequently
log to the ns_log, this should not cause any performance issues.

Change-Id: I1a1fbf2cea9ce4ef300846c86eb2f714c26b5710
Reviewed-on: http://review.couchbase.org/24910
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>