MB-100: provide api for testrunner for messing with memcached
I.e. because it currently relies on ns_port_sup which is now dead.

This also creates special erlang module explicitly dedicated for all
kinds of tricks testrunner or any other testsuite wishes to perform on
guts of ns_server. But in a clean, controlled and explicitly
ns_server-supported way.

Change-Id: I8f3597f1e9f681f6df481be55f8f2b9f3f70f570
Reviewed-on: http://review.couchbase.org/25475
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>