MB-100: use erlang:now for timer time
I've learned that erlang:now is not what I thought it is. It is
actually monotonic clock that's in case of real time clock jump is
slowly and gradually adjusted (by tweaking it's pace at most 1% of
normal clock rate) to match real time clock.

Apparently (but I haven't verified in code) receive timeouts are also
adjusted accordingly. But even if not main point still holds.

erlang:now with it's behavior is clearly more suitable for
timer. I.e. it is indeed tolerant to real time clock jumps in either
direction. While os:timestamp is clearly not.

Change-Id: I05c3e6987c2ec27df0b1d3f122d8c34575a2f1c8
Reviewed-on: http://review.couchbase.org/25575
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>