MB-100: start babysitter node with nocookie
I've seen that if I don't specify cookie then it'll generate some
sensible and randomly-looking cookie itself (but interstingly, not on
windows). So I decided to allow it to use it's own generated cookie.

But I've just found that something (presumably erlang itself) is
saving that generated cookie to /opt/couchbase/.erlang.cookie. That's
world-readable location (at least as of now) and thus is not really
secure.

Starting with nocookie will cause babysitter to generate cookie via
our normal cookie generation code.

Change-Id: Ic5e790d9ddbc4f077f210bf545985081e634695d
Reviewed-on: http://review.couchbase.org/25576
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>