MB-8013 Detailed rebalance progress backend.
We have new service (ns_rebalance_observer) which gets started
whenever rebalance is initiated. The service subscribes to
master_activity_events to monitor the proceeding of rebalance. When
rebalance for certain bucket starts, it makes initial estimate of
number of documents to be transferred. This estimate may be far
off. The estimates are refined when we create tap connections for
replication building. Additionally we have a process that periodically
loops through all current moves and updates the number of documents
that are left to transfer.

Change-Id: I6a3cfef9e5b1da4bedc5643c1303315a1cc71017
Reviewed-on: http://review.couchbase.org/25908
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>