MB-8189 Start child_erlang in a intermediate supervisor.
The potential problem with current approach is that ep-engine depends
on mccouch. So if for some reason ns_babysitter_sup is trying to
shutdown ns_child_ports_sup while child VM crashes, it will fail
because of this dependency. And at the same time child VM won't be
restarted because supervisor is waiting for ns_child_ports_sup to
terminate. With intermediate supervisor child VM will be restarted and
ep-engine will be able to shutdown.

Change-Id: I231ab390ab8de10dbdde61801769f80eae5bb069
Reviewed-on: http://review.couchbase.org/26089
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>