MB-100 Don't log ns_rebalance_observer state.
In MB-8237 I saw logs being rotated after 1.5 hours of rebalance
because of this. I considered just increasing the interval but it
seems that it's better to just add a little bit more logging in
critical places. In this case the amount of data we log is linear in
number of moves as opposed to linear in time with previous approach.

Change-Id: I77c9d8ea84561feaf547a534729a352544a6c0c2
Reviewed-on: http://review.couchbase.org/26232
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>