MB-8263 Start ns_bucket_{worker,sup} in a dedicated supervisor.
With previous approach if both ns_bucket_worker and ns_bucket_sup
crash at about the same time, it was possible to lose the work that
would start single_bucket_sup's. To prevent this both ns_bucket_worker
and ns_bucket_sup are now started in a dedicated supervisor with
one_for_all strategy. So that if any of them crashes, the other gets
restarted and correctly reinitialized.

Change-Id: I288676eed3f016c1708d8f79b295a1f109c83835
Reviewed-on: http://review.couchbase.org/26331
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>