MB-8309: bump ebucketmigrator's TERMINATE_TIMEOUT
So that it more or less matches other recently bumped timeouts.

As part of investigating MB-8309 we saw that we were hitting this
timeouts in initial phase of old-style vbucket filter change.

Our thinking is this timeouts is indeed needed but given other
timeouts were bumped and this one did not, we just need to bump this
one.

Change-Id: Ia8e02a211a094cca87ac866d341d425277247f88
Reviewed-on: http://review.couchbase.org/26416
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>