MB-8328 Don't set ERL_MAX_PORTS.
The difference in babysitter memory consumption with lowered and
default values is negligible. And our attempt to unset it for child VM
was incorrect. It would cause the VM (at least on r14b04 release) to
set the port limit to 1024. And the latter is much less than what we
really want.

Change-Id: Ia29aee37711f5c2c59a4ca1932746fc41c94b8ed
Reviewed-on: http://review.couchbase.org/26501
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>