MB-8352: take index path from node details in 1st wizard step
I've also removed incorrect escapeHTML in call to #val. It is not
needed because val is actually doing it via DOM properties and that
doesn't need any escaping.

Another change is to compute free space for db and index path based on
mount point disk free space data we have.

Change-Id: I8dc33863c2a4764e08856e9237438bca9798b122
Reviewed-on: http://review.couchbase.org/26609
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>