MB-8404: prevent replacing 127.0.0.1 in engageCluster reply
That replacement is usually needed for single-node clusters where node
name is 127.0.0.1. And where we cannot send 127.0.0.1 because that
will cause client to try to reach memcached (or moxi) at wrong
address.

But clearly this renaming is not needed in engageCluster which is
intra-cluster API. And also this renaming was actual reason why our
normally very strict pre-join checks failed to detect badness of
trying to join 127.0.0.1.

Change-Id: I34bb56d0018ae626adb8b79f5ac603ef3c434569
Reviewed-on: http://review.couchbase.org/26738
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>