MB-8533 make statistics work on node that is pending rebalance
if the node is in pending rebalance stage, then it is
not mapped to the bucket yet and all the servers that
provide statistical info are not started

so in this case we need to call other nodes to get
the statistics descriptions.

also the code uses node() as the first node to
get stat information from and if the first call
fails it will display empty stats.

the fix checks if the node() is one of the live nodes
for the bucket, and if not it will pick up the first
live node and perform the first call against it.

since the views signatures cannot be obtained from
the 2.0 nodes and at this time there's no compat mode
between 2.2 and 2.0 I decided that we can scrifice
"View Stats" for the added and non rebalanced nodes.

so in the currect solution View Stats will be hidden
for all the nodes that do not have the bucket yet.

alternative to this will be trying to obtain the info
from the other nodes in the cluster crashing
capi_set_view_manager on the older nodes.

another alternative is to create a separate gen_server
for obtaining the signatures. in this case the older
nodes will return noproc, and if the cluster contains
at least one 2.2 node that has the bucket we will
be able to populate Views Stats. But I decided that it
will be an overkill for such unimportant corner case.

Change-Id: Iefb41122a2696069278c112c15ce2c28f7570bdd
Reviewed-on: http://review.couchbase.org/27669
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>