MB-8828: better error handling in XMem mode XDCR
Today if XMem replicator received some error msgs which
it did not expect to see, it would crash. This behavior is
inherited from CAPI mode XDCR. For XMem mode, we should be
able to do better than that since now we send requests
to remote on a per-mutation basis. Instead of crashing
replicator, this commit will absorb and dump error msgs
from remote memacached into logs.

Change-Id: I950eb8cada03d7742108b8b140318c7ebb4c775d
Reviewed-on: http://review.couchbase.org/28234
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>