MB-8905: modify resolution algorithm for deletion in XMEM
The fix of MB-8825 at ep_engine side modified the resolution
algorithm for deletion, that is, we only compare Seqno and CAS
for del mutation in delWithMeta(). Because XDCR provides an
option to do local conflict resolution for backward compati-
bility, we also need to change the XDCR side conflict resolution
algorithm accordingly, to be consistent with ep_engine side
change.  The bottom line is, resolution algorithm should be the
same regardless of the location where we do conflict resolution.

Change-Id: I0a107b743b3b5b4f0ed9889c279ab0ebe1fa9c23
Reviewed-on: http://review.couchbase.org/28405
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>