MB-5904 Filter out admin_pass and pass in cbcollect_info.
Ideally, I would want all of this to be done in dump-guts. And for it
to use the same ns_config_log:sanitize function that is used by the
server. But sharing code between server and dump-guts script is not
very easy mainly because we our installation layout differs between
supported platforms. So at for now I decided to keep to our current
approach even though it requires code duplication.

Change-Id: Iac3175da07859d5f5ae86b8de31c5fb18b6b6105
Reviewed-on: http://review.couchbase.org/28413
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>