MB-8930: avoid crash replicator when timeout in getMeta
It is hard to expect when remote memcached will return the key
metadta that source XMem requests. Today we crash replicators if
the getMeta() to remote timed out or source receives some unexpected
memcached errors. A better behavior is that we send doc to remote
optimistically if we are unable to fetch the metadata from remote
memcached. The del/setWithMeta at remote cluster will do resolution
for us.

Also, we make separate stat for timeout from other errors.

Change-Id: Idddc07e287a5d56a4410af10c56b81fa39daf800
Reviewed-on: http://review.couchbase.org/28483
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Artem Stemkovski <artem@couchbase.com>