MB-8958: identify 2.2.0 "for xdcr" bucket list requests
In future we might plan to evolve bucket details response to remove
saslPassword field particularly. But that field is used by xdcr (xmem
mode). In order to be able to remove that field without breaking xdcr
backwards compatibility we'll start passing unused http query option.
So that future version can identify 2.2.0 requests and send it
saslPassword field as expected.

Change-Id: I0eddfef807310a2066348a97f5b3a3da785b2a5d
Reviewed-on: http://review.couchbase.org/28540
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>