MB-100: use rep_details to check replication mode
Today in vb replicator we use xmem_srv process to check the mode
of replication, which is correct in most cases except that if it
crashed, the terminate function of vb replicator will not be able
to know which mode the vb replicator is using. Actually we have
already encoded replication mode info in rep_details, thus we
should use it instead of relying on xmem_srv process.

Change-Id: I22816e3cea5f478f699cd47940a67688d76114e5
Reviewed-on: http://review.couchbase.org/28561
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>