MB-100: do not bother ns_orchestrator for recovery_status
Unless recovery is actually running.

In most common case of not being in recovery mode, we'll save remote
call for tasks request.

Because it's important for us to make sure recovery is visible in
/tasks right after start-recovery request, we also ensure that node
that handles start-recovery request (and thus performs rpc to
orchestrator) is in list of nodes we synchronize config to.

Change-Id: I67183eec44561deb096d5848549d5315420f7e24
Reviewed-on: http://review.couchbase.org/28776
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>