MB-6203 Fix race in waiting for bucket deletion on nodes.
We first delete bucket from config and then subscribe to
buckets_events. So it's possible to lose some of the events. The fix
is to explicitly filter out nodes for which predicate already holds
after the subscription is created. So even if some events are actually
lost, wait_for_nodes_loop won't senselessly wait for events about the
nodes that already satisfy the predicate.

Change-Id: I7881e17c3e5aace8a5b4ef496a2445d891d17569
Reviewed-on: http://review.couchbase.org/28869
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>