MB-8665: set TCP_NODELAY on port 8091 sockets
We've found that otherwise with http/1.1 keepalive enabled last
tcp segment of response is delayed as per Nagle's algorithm.

Ideally we'd use TCP_CORK and only pulse TCP_NODELAY after responses,
but erlang doesn't expose TCP_CORK to us.

With this setting we're sending response headers and body in separate
packages. But even without disabled Nagle's algorithm this is
happening too. So we're not making it worse.

Change-Id: I53bc2da462d15765eecc065689fd2fd8bc0984a2
Reviewed-on: http://review.couchbase.org/28952
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>