MB-8665: implemented caching of pool details request
We're putting it into menelaus_web_cache (keyed by ISAdmin and
LocalAddr that are inputs to pool info building).  We expire it after
one second because people might be polling it for certain stats we
unfortunately expose there.

We also invalidate cache on every config change. So that things like
changing certain settings are immediately reflected in subsequent
/pools/default output.

Also because UI depends on freshness of pool details (and up-to-date
buckets and tasks versions appended to corresponding urls advertised
from pool details) we only use (and update) cache for non-UI requests.

Change-Id: Id2115c81830567f0ef7bb2c937f4d17563c67cc4
Reviewed-on: http://review.couchbase.org/28958
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>