MB-8039: don't ping tap connections during janitor runs
The issue with pinging is that if node is down it may take couple
minutes for us to detect that. And while pinging we had node's
janitor_agent blocked.

So when any node was down that caused essentially all other node's
janitor agents to be blocked on pinging tap connection from that node
preventing failover actions.

That pinging was originally added to fight rebalance failure right
after node is rebooted. In which case we might have tap replications
from that node, not really aware that they're dead.

In order to fight that issue we're now simply specially detecting
crash from vbucket filter change caused by dead connection. If that
happens we simply establish fresh tap connection (which will kill
named tap producer first).

Change-Id: I5b5c4ddc5c12830dcca596ce9c9c055324332afe
Reviewed-on: http://review.couchbase.org/29051
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>