MB-100 Use misc:ensure_dir in path_config:component_path.
Because filelib:ensure_dir only ensures that all the parent
directories for certain path exist. So if the path passed to it does
not have trailing slash (which is not the case when the server is run
using cluster_run but is the case for our packages), it won't try to
create the directory with corresponding name. Our own version of
ensure_dir, on the other hand, will do this for us and it will also
ensure that we're able to write to this directory.

Change-Id: I7f77032a9e45c6d999701fe2ed65e18247dc1a82
Reviewed-on: http://review.couchbase.org/29077
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>