MB-9096: killed ns_online_config_upgrader
Which is not needed anymore.

I.e. majority of this code is handling upgrade from pre-2.0 to 2.0.

A bit more controversial piece is removal of xdcr settings which were
meant to be upgraded as part of upgrade to 2.2. But apparently it was
never working.

If we ever need similar code, then returning this module is going to
be easy

Change-Id: Icf682f560da0732c42680bf992d362d7e1882508
Reviewed-on: http://review.couchbase.org/29318
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>