bp: MB-9209: fix unbounded worker stats dictionary
We incorrectly used {<pid>, <unique ref>} from worker processes as
the key used to store worker stats in the vb replicator state. The
problem is each time a worker returns a different <unique ref> that
causes a new entry is created when worker reports stats to the vb
replicator. Overtime, the stats table will increase unboundedly,
causing memory leak. This fix uses a unique worker id to identify
each worker process and use that id as the key of worker stats table.
Thus the number of entries of vb stats table will be bounded by the
number of worker processes (default is 4).

Change-Id: Ic5cc7b1a9df4a0d587ce786f719414ee25c89246
[alk@tut.by: re-gerriting against 2.1.0]
Signed-off-by: Aliaksey Kandratsenka <alk@tut.by>
Reviewed-on: http://review.couchbase.org/29912
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>