bp:MB-9209: fix unbounded worker stats dictionary
We incorrectly used {<pid>, <unique ref>} from worker processes as
the key used to store worker stats in the vb replicator state. The
problem is each time a worker returns a different <unique ref> that
causes a new entry is created when worker reports stats to the vb
replicator. Overtime, the stats table will increase unboundedly,
causing memory leak. This fix uses a unique worker id to identify
each worker process and use that id as the key of worker stats table.
Thus the number of entries of vb stats table will be bounded by the
number of worker processes (default is 4).

Change-Id: I4c02fba251e234918c3d1b96931d0cc1acf6e4d5
Orig-Change-Id: Iac9461cb2ddad37e6ee0edce4ef7d87e0ceef1e3
Orig-Reviewed-on: http://review.couchbase.org/29907
Reviewed-on: http://review.couchbase.org/30549
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>