MB-9460: dont crash rebalance in older compat modes
Recent code change blindly assumed that server groups is always going
to be present. But this is not the case as we're only creating that
config key as part of upgrading compat mode to 2.5 (i.e. when all
nodes are 2.5+). So in order to support mixed-version rebalance with
older versions we have to anticipate lack of this key

Change-Id: Ibe8c34f1e147668c818b267678c5709311f3ee72
Reviewed-on: http://review.couchbase.org/30104
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>