MB-100: removed manual ++ optimization
It turns out ++ BIF is doing it more or less in same way. I.e. see
erl_bif_lists.c:append_2. The only difference is that it validates
first argument for being list (and computes it's length) prior to
examining second arg emptyness. But because that's C code and because
we don't expect our lists to be huge, I believe leaving that
optimization to C is benefical. And simpler code is a win in itself.

Change-Id: If45b11978e778fd37247de9208789240419a3d96
Reviewed-on: http://review.couchbase.org/30217
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>