MB-100: mark expected failed cells test as expected failing
I.e. because I've just spent 10 minutes bisecting for this
failure. And predictably, ended up discovering that I've created this
failing unit test long time ago planning to fix cells to make it pass.

Given plans for cells, it'll clearly not going to be fixed. So we need
to either remove that test or just mark it as expect passing to
prevent folks from wasting time trying to fix it.

Change-Id: I14cfb5524d6b182f29f0396430322e36e81e60ab
Reviewed-on: http://review.couchbase.org/30299
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>