MB-9589: handle incomplete cluster leave
When cluster leave function fails we can end up in a state where node
remains in a limbo state forever.

New code places marker file on filesystem to indicate that leave is
happening. With that, if ns_cluster dies during leave cluster
procedure, it'll detect incomplete leave in init and complete it (or
fail while doing it).

Change-Id: I41fa757a2527de8e95cfdad95029b06216f7880b
Reviewed-on: http://review.couchbase.org/30413
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>