MB-9652: avoid restarting net_kernel unless absolutely necessary
One cause of issues in the ticket appears to be net_kernel
shutdowns. In most common case of not assigning hostnames this is not
actually required. So we don't touch it in this cases.

It certainly doesn't address root of the problem (which is still
unknown at this time but is believed to be somewhere within erlang
itself). But at least this commit should buy us enough time to
investigate and fix/workaround it.

Change-Id: I246f7180e6c49552f338cfefdf746775c48b6778
Reviewed-on: http://review.couchbase.org/30873
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>