MB-9725 Handle tags when looking for matching renamings.
In the presence of tags it's not enough to find nodes that can be
swapped. It's also important to check that tags after renaming are
compatible with currently set tags.

The tags themselves are now ignored when a past map is considered for
compatibility. And then matching_renaming_with_tags will just ignore
the map if it's incompatible.

Change-Id: Iea35215ad0558cf1de26731576e6057200ecff32
Reviewed-on: http://review.couchbase.org/31063
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>