MB-9725 Ignore map options when updating vbucket map history.
When putting some vbucket map to the history the only thing we are
interested is actual balancedness of the vbucket map. We are not
interested if map options are the same or different from current
ones. As long as the map is balanced, it can be used in conjunction
with its own options. For example, when bucket is created and new map
is generated, previous version of ns_rebalancer:unbalanced would
consider this map unbalanced just because bucket config doesn't have
map_opts_hash in it. Same is true for the upgrade case where we put
current buckets' maps to the history (though ns_rebalancer:unbalanced
had a special case for this).

So I factored out ns_rebalancer:map_options_changed function out of
ns_rebalancer:unbalanced. And I call it only in places where it
actually matters.

Change-Id: I11d67ec84a335f544b5d5c2b93beb78453a19419
Reviewed-on: http://review.couchbase.org/31080
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>