MB-9663: consume src_db_update msg when rep is in init
Due to an earlier change, a vb replicator may receive a src_db
_update msg when it is still in initialization. But vb replicator
does not know handle it, leading to vb replicator crashing at the
function_clause error. This commit adds the handle function, which
basically does nothing but consuming the msg, because replicator
will send itself a src_db_update msg after init is over to check
if any update to replicate.

Change-Id: Ie7d39ec7b5a0a3511991b23a8f2ee8f2215f8e2c
Reviewed-on: http://review.couchbase.org/31086
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>