MB-9745: delay second restart as little as possible
Rather than full configured restart delay. Otherwise with default
setting of 30 seconds we will only attempt restarts at intervals that
are multiple of 30 seconds.

This is achieved via supervisor2 feature of restart delay 0. Which is
implemented to wait until oldest restart expires.

This is then used to pass 0 as delay for second restart attempt.

Change-Id: Ib046de2e76d9c3c6f2865ee85b5eb53f903f0150
Reviewed-on: http://review.couchbase.org/31762
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>