MB-9343 wait for data move implementation
The first event that ns_server needs to see take place
is for the new VBucket to be close to being up to date.
NS_Server will be able to find out about this event
by polling stats "upr-vbtakeover <vbid> <connection name>"
looking at the value backfillRemaining on the
producer node.

This stat will provide an estimate of how much more data
needs to be sent over the given stream for a given vbucket.
If there are less than 1000 items that still need to be sent
the NS_Server can assume that the new VBucket is almost
up to date with respect to the old VBucket.

Change-Id: I0e8abb75ea4c5b92a6f38bf8cd6b8e7bcce8f6f5
Reviewed-on: http://review.couchbase.org/33197
Tested-by: Artem Stemkovski <artem@couchbase.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>