Update vbmap tool.
Main change compared to previous version is faster 3rd phase which is
generation of actual vbucket map from matrix R. It's not an order of
magnitude difference, but "go test" now takes about 90 seconds on my
machine versus 220 using the old code.

In addition to that, when we try to minimize pairwise occurrences of
nodes over vbucket map chains, we do it for the entire vbucket map and
not only for the chains that have the same master node as it used to be.

Change-Id: I9fa03c5c7e0b630f5d45e6f26d44be2eff78e4e9
Reviewed-on: http://review.couchbase.org/32625
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>