MB-9883: control batching of memcached replies in proxy
We now have slightly different proxy protocol that allows proxy on
remote node to be aware of memcached batches.

For that every batch now starts with 32-bit batch size in bytes
followed by 32-bit batch size in requests. Downstream proxy is using
that to accumulate proper count of replies before sending all of them
via single ssl:send call.

We have to break abstractions a bit to support that. Particularly
proxied_memcached_clients_pool has to be aware of batching.

Change-Id: I960d50b884ede1a96b142acc3be268224618ec60
Reviewed-on: http://review.couchbase.org/32704
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>