MB-9988: bump had_backfill timeout to infinity
We have a general policy of having infinite erlang timeouts for
everything related to rebalance. The logic is as soon as rebalance
stop works without relying on those timeouts, we're fine. And stop
doesn't rely on timeouts because those calls are wrapped with
spawn_and_wait which was built specifically to be able to abort
something when rebalance is stopped.

Change-Id: Ie49d24bb19b87669b971a94d2612b5941fec75ae
Reviewed-on: http://review.couchbase.org/32749
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>