Restore pretty-printing of stats.
Stats will still be returned to us by old nodes. Therefore with current
code they would be printed in the very end of handling diag by
do_continue_handling_per_node_just_diag in a bulk. This might lead to
high memory usage since pretty-printer is a memory hog. So it's safer to
restore the old code that handles stats in a special way.

Change-Id: I4e30e384bb2772a302505450adc479944d4ddb2b
Reviewed-on: http://review.couchbase.org/33060
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>