Hide normal port terminations from user visible log correctly.
Since ale:info is not actually a function, but something that gets
transformed to an actual function call by ale_transform parse
transform, it's as smart as a parse transform is, which is not. So
it actually doesn't understand when logger name is passed through a
variable. ale:log is used instead which is smarter in this respect.

Also instead of using different logger, now I'm just adjusting
message's log level.

Change-Id: I30a8b0e40a1a09f7dcd8c78537f496afa4bdffe3
Reviewed-on: http://review.couchbase.org/33087
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>